requirements.txt
py-clob-client>=0.17.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-dotenv>=1.0.0
ccxt>=4.0.0
requests>=2.31.0
websockets>=12.0
schedule>=1.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
